extracting debuginfo from powerFxengine to loggerhelper
