+++ src/Microsoft.PowerApps.TestEngine/Helpers/ILoggerHelper.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Microsoft.PowerApps.TestEngine.Helpers
+{
+    public interface ILoggerHelper
+    {
+        /// <summary>
+        /// Log debug info
+        /// </summary>
+        public void DebugInfo();
+    }
+}
